+++ CHANGELOG.md
@@ -5,11 +5,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.0.0-beta.XX] - 2019-XX-XX
 ### Added
 - Added Search Results overlay
-- Added Search Engine - PostgreSQL
 - Added Search Engine - DB Basic
+- Added Search Engine - DB PostgreSQL
+- Added Search Engine - Azure Search
+- Added Search Engine - AWS CloudSearch
 - Added Git changes processing (add/modify/delete)
 - Added Storage last sync date in status panel
 - Added Dev Flags
+- Added HTTPS server option
 - Added HTTP to HTTPS redirect server option
 
 ### Fixed